refactor: migrate project from Next.js to Vite and update API endpoint logic for flexible environment configuration
