--- next.config.js
@@ -1,10 +1,0 @@
-/**
- * Run `build` or `dev` with `SKIP_ENV_VALIDATION` to skip env validation. This is especially useful
- * for Docker builds.
- */
-import "./src/env.js";
-
-/** @type {import("next").NextConfig} */
-const config = {output: 'standalone', reactStrictMode: true,};
-
-export default config;
