+++ package.json
@@ -4,24 +4,19 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "next build",
-    "check": "next lint && tsc --noEmit",
-    "dev": "next dev --turbo",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "format:check": "prettier --check \"**/*.{ts,tsx,js,jsx,mdx}\" --cache",
     "format:write": "prettier --write \"**/*.{ts,tsx,js,jsx,mdx}\" --cache",
-    "lint": "next lint",
-    "lint:fix": "next lint --fix",
-    "preview": "next build && next start",
-    "start": "next start --port 3001",
+    "lint": "eslint .",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "@headlessui/react": "^2.2.9",
     "@heroicons/react": "^2.2.0",
-    "@t3-oss/env-nextjs": "^0.13.10",
     "d3-scale": "^4.0.2",
     "lucide-react": "^1.11.0",
-    "next": "^16.2.4",
     "plotly.js": "^3.3.1",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
@@ -31,19 +26,21 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.5",
     "@tailwindcss/postcss": "^4.2.4",
+    "@tailwindcss/vite": "^4.2.4",
     "@types/d3-scale": "^4.0.9",
     "@types/node": "^25.6.0",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "@types/react-plotly.js": "^2.6.4",
+    "@vitejs/plugin-react": "^4.3.4",
     "eslint": "^10.2.1",
-    "eslint-config-next": "^16.1.1",
     "postcss": "^8.5.12",
     "prettier": "^3.8.3",
     "prettier-plugin-tailwindcss": "^0.7.2",
     "tailwindcss": "^4.2.4",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.59.0"
+    "typescript-eslint": "^8.59.0",
+    "vite": "^6.2.0"
   },
   "ct3aMetadata": {
     "initVersion": "7.39.3"
