--- src/app/layout.tsx
@@ -1,25 +1,0 @@
-import "~/styles/globals.css";
-
-import { type Metadata } from "next";
-import { Geist } from "next/font/google";
-
-export const metadata: Metadata = {
-  title: "Malware Visualizer",
-  description: "Generated by create-t3-app",
-  icons: [{ rel: "icon", url: "/favicon.ico" }],
-};
-
-const geist = Geist({
-  subsets: ["latin"],
-  variable: "--font-geist-sans",
-});
-
-export default function RootLayout({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <html lang="en" className={`${geist.variable}`}>
-      <body>{children}</body>
-    </html>
-  );
-}
