+++ src/hooks/useMalwareData.ts
@@ -1,4 +1,4 @@
-"use client";
+
 
 import { useState, useTransition, useCallback } from "react"; // 1. Import useCallback
 import type { RawApiResponse, MalwareMetadata } from "~/interfaces/malware";
@@ -40,37 +40,36 @@ export function useMalwareData() {
 
           setStatusMessage(`Fetching data from collection: "${collection}"...`);
 
-          const apiUrl = new URL("/api/fastapi-data", window.location.origin);
-          apiUrl.searchParams.set("collection_name", collection);
-          apiUrl.searchParams.set("apply_dr", String(applyDR));
-          apiUrl.searchParams.set("dr_method", drMethod);
-          apiUrl.searchParams.set("n_components", String(nComponents));
+          const apiBase = import.meta.env.DEV ? "/api" : (import.meta.env.VITE_FASTAPI_URL || "");
+          
+          const finalUrl = (!import.meta.env.DEV && import.meta.env.VITE_FASTAPI_URL?.startsWith('http')) 
+            ? new URL("/query_weaviate", import.meta.env.VITE_FASTAPI_URL)
+            : new URL(`${apiBase}/query_weaviate`.replace(/\/+/g, '/'), window.location.origin);
 
-          const response = await fetch(apiUrl.toString());
+          finalUrl.searchParams.set("collection_name", collection);
+          finalUrl.searchParams.set("apply_dr", String(applyDR));
+          finalUrl.searchParams.set("dr_method", drMethod);
+          finalUrl.searchParams.set("n_components", String(nComponents));
+          finalUrl.searchParams.set("query", "");
+
+          const response = await fetch(finalUrl.toString());
 
           if (!response.ok) {
-            const errorBody = (await response.json()) as NextApiResponseError;
-            throw new Error(
-              errorBody.error || `HTTP error! status: ${response.status}`,
-            );
+            const errorText = await response.text();
+            throw new Error(`HTTP error! status: ${response.status}, message: ${errorText}`);
           }
 
-          const result = (await response.json()) as {
-            success: boolean;
-            data?: RawApiResponse;
-            error?: string;
-            message?: string;
-          };
+          const result = (await response.json()) as RawApiResponse;
 
-          if (result.success && result.data) {
-            if (!result.data.results || !Array.isArray(result.data.results)) {
+          if (result) {
+            if (!result.results || !Array.isArray(result.results)) {
               throw new Error(
                 "API response is malformed: missing 'results' array.",
               );
             }
             const embeddings: number[][] = [];
             const metadataList: MalwareMetadata[] = [];
-            for (const item of result.data.results) {
+            for (const item of result.results) {
               if (item.embedding && item.metadata) {
                 embeddings.push(item.embedding);
                 metadataList.push(item.metadata);
@@ -80,17 +79,16 @@ export function useMalwareData() {
               data: embeddings,
               metadata: metadataList,
               shape: [embeddings.length, embeddings[0]?.length ?? 0],
-              pacmap_applied: result.data.pacmap_applied ?? applyDR,
-              collection_name: result.data.collection_name ?? collection,
+              pacmap_applied: result.pacmap_applied ?? applyDR,
+              collection_name: result.collection_name ?? collection,
               message:
-                result.data.message ??
                 result.message ??
                 "Data processed and loaded successfully.",
             };
             setData(processedData);
             setStatusMessage(processedData.message);
           } else {
-            throw new Error(result.error ?? "An unknown API error occurred.");
+            throw new Error("An unknown API error occurred.");
           }
         } catch (error) {
           console.error("Error in useMalwareData hook:", error);
